socket-chat-server/requirements.txt
eventlet==0.31.0
Flask==2.0.1
Flask-SocketIO==5.1.0
greenlet==1.1.0
python-decouple==3.4